refactor: move Github Button to a component
