+++ example-nextjs/components/Navbar.tsx
